Cases now get correct publication number and moved cases data hooks to appropriate location
