+++ package.json
@@ -8,7 +8,7 @@
     "lint": "nx lint",
     "run-xroad-proxy": "./scripts/run-xroad-proxy.sh",
     "license-audit": "license-checker --failOn 'GPL; AGPL; UNLICENSED' --summary",
-    "openapi-generator": "openapi-generator generate -g typescript-fetch --additional-properties=typescriptThreePlus=true",
+    "openapi-generator": "openapi-generator generate -g typescript-fetch --additional-properties=typescriptThreePlus=true --skip-validate-spec",
     "codegen": "node scripts/codegen.js",
     "compile-strings:is": "formatjs compile ./apps/official-journal-web/src/i18n/strings/is.json --ast --out-file ./apps/official-journal-web/src/i18n/strings/is-compiled.json",
     "compile-strings": "concurrently \"pnpm:compile-strings:*\"",
